Add fix for spark_df flattener
On this Commit all fixes for spark_df have been applied.

There was an error with 'replace_dots' param that was solved.

Also there is an improvement on the case translator, where it could be set to None and leave columns and dataframe names as default.

The replace_dots param was replaced by the replace_string variable, it allows to separate nested data by any special character that user wants.

Finally there is a new option called "remove_special_characters" used on the CaseTranslator to enable the special characters removal from the column and dataframe names.